+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
+## Unreleased
+### Added
+- Price change reasons are now passed as an array
+
 ## [3.2.0]
 ### Changed
 - Return status with QuickTravel::AdapterError